+++ apps/roam/src/utils/initializeObserversAndListeners.ts
@@ -40,10 +40,7 @@ import {
   removeTextSelectionPopup,
   findBlockElementFromSelection,
 } from "~/utils/renderTextSelectionPopup";
-import { renderNodeTagPopup, removeNodeTagPopup } from "./renderNodeTagPopup";
-import getTextByBlockUid from "roamjs-components/queries/getTextByBlockUid";
-import { renderCreateNodeDialog } from "~/components/CreateNodeDialog";
-import getUids from "roamjs-components/dom/getUids";
+import { renderNodeTagPopupButton } from "./renderNodeTagPopup";
 
 const debounce = (fn: () => void, delay = 250) => {
   let timeout: number;
@@ -65,7 +62,6 @@ export const initObservers = async ({
     nodeMenuTriggerListener: EventListener;
     discourseNodeSearchTriggerListener: EventListener;
     nodeCreationPopoverListener: EventListener;
-    nodeTagHoverListener: EventListener;
   };
 }> => {
   const pageTitleObserver = createHTMLObserver({
@@ -96,6 +92,14 @@ export const initObservers = async ({
   const queryBlockObserver = createButtonObserver({
     attribute: "query-block",
     render: (b) => renderQueryBlock(b, onloadArgs),
+  });
+
+  const nodeTagPopupButtonObserver = createHTMLObserver({
+    className: "rm-page-ref--tag",
+    tag: "SPAN",
+    callback: (s: HTMLSpanElement) => {
+      renderNodeTagPopupButton(s, onloadArgs.extensionAPI);
+    },
   });
 
   const pageActionListener = ((
@@ -298,52 +302,6 @@ export const initObservers = async ({
     }
   }, 150);
 
-  const nodeTagHoverListener = (e: Event) => {
-    const target = e.target as HTMLElement | null;
-    if (!target || !target.classList?.contains("rm-page-ref--tag")) return;
-
-    const textContent = target.textContent?.trim() || "";
-    const tagAttr = target.getAttribute("data-tag") || textContent;
-    const tag = tagAttr.replace(/^#/, "").toLowerCase();
-    const discourseNodes = getDiscourseNodes();
-    const discourseTagSet = new Set(
-      discourseNodes.map((n) => n.tag?.toLowerCase()).filter(Boolean),
-    );
-
-    if (!discourseTagSet.has(tag)) return;
-
-    const matchedNode = discourseNodes.find(
-      (n) => n.tag?.toLowerCase() === tag,
-    );
-
-    if (!matchedNode) return;
-
-    const blockInputElement = (e.target as HTMLElement).closest(
-      ".rm-block__input",
-    );
-    const blockUid = blockInputElement
-      ? getUids(blockInputElement as HTMLDivElement).blockUid
-      : undefined;
-
-    const rawBlockText = blockUid ? getTextByBlockUid(blockUid) : "";
-    const cleanedBlockText = rawBlockText.replace(textContent, "").trim();
-
-    renderNodeTagPopup({
-      tagElement: target,
-      label: `+ Create ${matchedNode.text}`,
-      onClick: () => {
-        removeNodeTagPopup();
-        renderCreateNodeDialog({
-          onClose: () => {},
-          defaultNodeTypeUid: matchedNode.type,
-          extensionAPI: onloadArgs.extensionAPI,
-          sourceBlockUid: blockUid,
-          initialTitle: cleanedBlockText,
-        });
-      },
-    });
-  };
-
   return {
     observers: [
       pageTitleObserver,
@@ -351,6 +309,7 @@ export const initObservers = async ({
       configPageObserver,
       linkedReferencesObserver,
       graphOverviewExportObserver,
+      nodeTagPopupButtonObserver,
     ].filter((o): o is MutationObserver => !!o),
     listeners: {
       pageActionListener,
@@ -358,7 +317,6 @@ export const initObservers = async ({
       nodeMenuTriggerListener,
       discourseNodeSearchTriggerListener,
       nodeCreationPopoverListener,
-      nodeTagHoverListener,
     },
   };
 };
